+++ CHANGELOG.md
@@ -13,6 +13,8 @@ project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 
+- feat: exposed `registry.registerCollector()` and `registry.collectors()` methods in TypeScript declaration
+
 ## [12.0.0] - 2020-02-20
 
 ### Breaking